Fix table tests
