+++ tests/jballerina-unit-test/src/test/resources/test-src/types/table/table_literal_syntax_negative2.bal
@@ -29,3 +29,7 @@ function testTableLiteralDataAndAddWithObject() returns (int) {
     int count = t1.count();
     return count;
 }
+
+function testEmptyTableCreateInvalid() {
+    table t1 = table{};
+}
