+++ tests/jballerina-unit-test/src/test/resources/test-src/types/table/table_test_helper.bal
@@ -63,4 +63,3 @@ function getSessionCount() returns (int) {
     checkpanic testDB.stop();
     return count;
 }
-
