+++ tests/jballerina-unit-test/src/test/resources/testng.xml
@@ -111,6 +111,7 @@
             <!--<package name="org.ballerinalang.test.record.*"/>-->
             <package name="org.ballerinalang.test.types.constant.*" />
             <package name="org.ballerinalang.test.types.bytetype.*"/>
+            <package name="org.ballerinalang.test.types.table.*"/>
             <!--<package name="org.ballerinalang.test.types.handle.*"/>-->
             <!--<package name="org.ballerinalang.test.balo.*"/>-->
             <package name="org.ballerinalang.test.expressions.typecast.*"/>
